move plugin classes to compat package
